Made compilation possible on windows

+++ src/api/vulkan/PipelineMap.cpp
@@ -71,9 +71,9 @@ VkPipelineLayout	PipelineMap::getLayout(void) {
 	_initLayout(_device, setLayouts, pushConstants);
 	VkPipelineLayoutCreateInfo	createInfo{};
 	createInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-	createInfo.setLayoutCount = setLayouts.size();
+	createInfo.setLayoutCount = static_cast<uint32_t>(setLayouts.size());
 	createInfo.pSetLayouts = setLayouts.data();
-	createInfo.pushConstantRangeCount = pushConstants.size();
+	createInfo.pushConstantRangeCount = static_cast<uint32_t>(pushConstants.size());
 	createInfo.pPushConstantRanges = pushConstants.data();
 	if (vkCreatePipelineLayout(_device.getLogical(), &createInfo,
 								nullptr, &_layout))
@@ -119,7 +119,7 @@ bool	PipelineMap::bindPipeline(const RenderingConfig &renderingConfig,
 		config.pipelineLayout = _layout;
 
 		config.renderingInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_RENDERING_CREATE_INFO;
-		config.renderingInfo.colorAttachmentCount = renderingConfig.colorFormats.size();
+		config.renderingInfo.colorAttachmentCount = static_cast<uint32_t>(renderingConfig.colorFormats.size());
 		config.renderingInfo.pColorAttachmentFormats = renderingConfig.colorFormats.data();
 		config.renderingInfo.depthAttachmentFormat = renderingConfig.depthFormat;
 		//TODO -> add it's own var in the config.
